# Improve IdentifierIndex firstIndexOf performance