Avoids doing two hasmap lookups and instead does a single one, thus improving performances (CPU) of rougthly 33%.

In applications reading lots of rows and not managing manually indexes (which is tricky, boiler plate and error prone),
IdentifierIndex::firstIndexOf is typically a CPU hotspots. In reactive applications it might even run on the driver event loop. As such, performances gains here are welcome.

## Before

```
Benchmark                             Mode  Cnt  Score   Error  Units
IdentifierIndexBench.complexGetFirst  avgt    5  0.046 ± 0.005  us/op
IdentifierIndexBench.simpleGetFirst   avgt    5  0.046 ± 0.002  us/op
```

## After

```
Benchmark                             Mode  Cnt  Score   Error  Units
IdentifierIndexBench.complexGetFirst    avgt    5  0.028 ± 0.002  us/op
IdentifierIndexBench.simpleGetFirst     avgt    5  0.030 ± 0.002  us/op
```

[Use ImmutableListMultimap within IdentifierIndex](https://github.com/datastax/java-driver/pull/1615/commits/1631b6d8c3e8a9cc46d5388382f40f527bb6a558)
[1631b6d](https://github.com/datastax/java-driver/pull/1615/commits/1631b6d8c3e8a9cc46d5388382f40f527bb6a558)

This unlocks massive performance gains upon reads, for a minor slow
down at instanciation time, which won't be felt by applications relying
on prepared statements.

## Before

```
Benchmark                               Mode  Cnt  Score   Error  Units
IdentifierIndexBench.complexAllIndices  avgt    5  0.007 ± 0.001  us/op
IdentifierIndexBench.complexGetFirst    avgt    5  0.026 ± 0.002  us/op
IdentifierIndexBench.createComplex      avgt    5  0.759 ± 0.059  us/op
IdentifierIndexBench.createSimple       avgt    5  0.427 ± 0.048  us/op
IdentifierIndexBench.simpleAllIndices   avgt    5  0.007 ± 0.001  us/op
IdentifierIndexBench.simpleGetFirst     avgt    5  0.027 ± 0.002  us/op
```

## After

```
Benchmark                               Mode  Cnt  Score    Error  Units
IdentifierIndexBench.complexAllIndices  avgt    5  0.004 ±  0.001  us/op
IdentifierIndexBench.complexGetFirst    avgt    5  0.005 ±  0.001  us/op
IdentifierIndexBench.createComplex      avgt    5  0.680 ±  0.020  us/op
IdentifierIndexBench.createSimple       avgt    5  0.538 ±  0.096  us/op
IdentifierIndexBench.simpleAllIndices   avgt    5  0.004 ±  0.001  us/op
IdentifierIndexBench.simpleGetFirst     avgt    5  0.005 ±  0.001  us/op
```