@chibenwa 2022-11-01
JMH code used here:

```
package com.datastax.oss.driver.internal.core.data;

import java.util.concurrent.TimeUnit;

import org.junit.Test;
import org.openjdk.jmh.annotations.Benchmark;
import org.openjdk.jmh.annotations.Mode;
import org.openjdk.jmh.infra.Blackhole;
import org.openjdk.jmh.runner.Runner;
import org.openjdk.jmh.runner.options.Options;
import org.openjdk.jmh.runner.options.OptionsBuilder;
import org.openjdk.jmh.runner.options.TimeValue;

import com.datastax.oss.driver.api.core.CqlIdentifier;
import com.datastax.oss.driver.shaded.guava.common.collect.ImmutableList;

public class IdentifierIndexBench {
    private static final CqlIdentifier ID_1 = CqlIdentifier.fromCql("id1");
    private static final CqlIdentifier ID_2 = CqlIdentifier.fromCql("id1");
    private static final CqlIdentifier ID_3 = CqlIdentifier.fromCql("id1");
    private static final IdentifierIndex SIMPLE_INDEX = new IdentifierIndex(ImmutableList.of(ID_1, ID_2, ID_3));
    private static final IdentifierIndex COMPEX_INDEX = new IdentifierIndex(ImmutableList.of(ID_1, ID_2, ID_3, ID_1, ID_2, ID_3));

    @Test
    public void launchBenchmark() throws Exception {
        Options opt = new OptionsBuilder()
            .include(this.getClass().getName() + ".*GetFirst")
            .mode (Mode.AverageTime)
            .timeUnit(TimeUnit.MICROSECONDS)
            .warmupTime(TimeValue.seconds(5))
            .warmupIterations(3)
            .measurementTime(TimeValue.seconds(2))
            .measurementIterations(5)
            .threads(1)
            .forks(1)
            .shouldFailOnError(true)
            .shouldDoGC(true)
            .build();

        new Runner(opt).run();
    }

    @Benchmark
    public void simpleGetFirst(Blackhole bh) {
        bh.consume(SIMPLE_INDEX.firstIndexOf(ID_1));
    }

    @Benchmark
    public void complexGetFirst(Blackhole bh) {
        bh.consume(COMPEX_INDEX.firstIndexOf(ID_1));
    }

    @Benchmark
    public void simpleAllIndices(Blackhole bh) {
        bh.consume(SIMPLE_INDEX.allIndicesOf(ID_1));
    }

    @Benchmark
    public void complexAllIndices(Blackhole bh) {
        bh.consume(COMPEX_INDEX.allIndicesOf(ID_1));
    }

    @Benchmark
    public void createSimple(Blackhole bh) {
        bh.consume(new IdentifierIndex((ImmutableList.of(ID_1, ID_2, ID_3))));
    }

    @Benchmark
    public void createComplex(Blackhole bh) {
        bh.consume(new IdentifierIndex((ImmutableList.of(ID_1, ID_2, ID_3, ID_1, ID_2, ID_3))));
    }
}
```