@chibenwa 2022-11-07
As expressed in https://github.com/datastax/java-driver/pull/1614#issuecomment-1305005793 this work is where big easy gains are. 

@absurdfarce any feedback on this work?