@absurdfarce 2022-11-07
@chibenwa I want to spend a little more time with this one but my strong inclination is to agree that there's a solid, easy win here.  I'll try to have this wrapped up in the next few days.