@adutra 2022-11-08
@chibenwa this is awesome.

Would it be worth exploring the idea of going one step further and precomputing the first indices?

Something like:

```java
public class IdentifierIndex {

  private final ListMultimap<CqlIdentifier, Integer> byId;
  private final ListMultimap<String, Integer> byCaseSensitiveName;
  private final ListMultimap<String, Integer> byCaseInsensitiveName;
  private final ImmutableMap<CqlIdentifier, Integer> firstById;
  private final ImmutableMap<String, Integer> firstByCaseSensitiveName;
  private final ImmutableMap<String, Integer> firstByCaseInsensitiveName;

  public IdentifierIndex(List<CqlIdentifier> ids) {
    ImmutableListMultimap.Builder<CqlIdentifier, Integer> byId = ImmutableListMultimap.builder();
    ImmutableListMultimap.Builder<String, Integer> byCaseSensitiveName = ImmutableListMultimap.builder();
    ImmutableListMultimap.Builder<String, Integer> byCaseInsensitiveName = ImmutableListMultimap.builder();

    int i = 0;
    for (CqlIdentifier id : ids) {
      byId.put(id, i);
      byCaseSensitiveName.put(id.asInternal(), i);
      byCaseInsensitiveName.put(id.asInternal().toLowerCase(Locale.ROOT), i);
      i += 1;
    }

    this.byId = byId.build();
    this.byCaseSensitiveName = byCaseSensitiveName.build();
    this.byCaseInsensitiveName = byCaseInsensitiveName.build();

    firstById = this.byId.entries()
        .stream()
        .collect(ImmutableMap.toImmutableMap(Entry::getKey, Entry::getValue, (i1, i2) -> i1));
    firstByCaseSensitiveName = this.byCaseSensitiveName.entries()
        .stream()
        .collect(ImmutableMap.toImmutableMap(Entry::getKey, Entry::getValue, (i1, i2) -> i1));
    firstByCaseInsensitiveName = this.byCaseInsensitiveName.entries()
        .stream()
        .collect(ImmutableMap.toImmutableMap(Entry::getKey, Entry::getValue, (i1, i2) -> i1));
  }

  /**
   * Returns all occurrences of a given name, given the matching rules described in {@link
   * AccessibleByName}.
   */
  public List<Integer> allIndicesOf(String name) {
    return Strings.isDoubleQuoted(name)
        ? byCaseSensitiveName.get(Strings.unDoubleQuote(name))
        : byCaseInsensitiveName.get(name.toLowerCase(Locale.ROOT));
  }

  /**
   * Returns the first occurrence of a given name, given the matching rules described in {@link
   * AccessibleByName}, or -1 if it's not in the list.
   */
  public int firstIndexOf(String name) {
    return Strings.isDoubleQuoted(name)
        ? firstByCaseSensitiveName.getOrDefault(Strings.unDoubleQuote(name), -1)
        : firstByCaseInsensitiveName.getOrDefault(name.toLowerCase(Locale.ROOT), -1);
  }

  /** Returns all occurrences of a given identifier. */
  public List<Integer> allIndicesOf(CqlIdentifier id) {
    return byId.get(id);
  }

  /** Returns the first occurrence of a given identifier, or -1 if it's not in the list. */
  public int firstIndexOf(CqlIdentifier id) {
    return firstById.getOrDefault(id, -1);
  }
}
```