@chibenwa 2022-11-08
Hello,

Thanks for the enthousiasm.

However I realized today that the massive gains are because ImmutableStuff from guava do a `==` call prior to hashing. The benchs written above leveraged that in the **massive** gains.

I wonder how this would thansfer to applications in the wild to be fairly honnest: if I use the same CqlIdentifier constant to build the statement AND when accessing the row would the driver preserve the `==` ? If so we have the gains, if not, that would be more limited.

Given that new knowledge I did do a new run of the benchmarks. Still meaningful. I included your proposal @adutra but it do not bring enhancements so I think it would not be worth the pain...

```
// 4.x
Benchmark                             Mode  Cnt  Score   Error  Units
IdentifierIndexBench.complexGetFirst  avgt    5  0.115 ± 0.005  us/op
IdentifierIndexBench.simpleGetFirst   avgt    5  0.108 ± 0.006  us/op

// PR-1 c150185 
Benchmark                             Mode  Cnt  Score   Error  Units
IdentifierIndexBench.complexGetFirst  avgt    5  0.097 ± 0.003  us/op
IdentifierIndexBench.simpleGetFirst   avgt    5  0.091 ± 0.014  us/op

// PR 1631b6d8c3e8a9cc46d5388382f40f527bb6a558
Benchmark                             Mode  Cnt  Score   Error  Units
IdentifierIndexBench.complexGetFirst  avgt    5  0.060 ± 0.001  us/op
IdentifierIndexBench.simpleGetFirst   avgt    5  0.073 ± 0.014  us/op

// @adutrat proposal
Benchmark                             Mode  Cnt  Score   Error  Units
IdentifierIndexBench.complexGetFirst  avgt    5  0.073 ± 0.011  us/op
IdentifierIndexBench.simpleGetFirst   avgt    5  0.068 ± 0.004  us/op
```